fix: satisfy DeepSeek usage lint rules
